Update action schema and refactor imports in tests
- Added "take-off" to the examples of the commandVerb in action.schema.json.
- Updated the description of the scope property for clarity.
- Introduced a new required_components property to specify component IDs needed for action discovery.
- Refactored imports in multiple test files to use a centralized index for SimpleEntityManager.
