+++ tests/common/index.js
@@ -1,0 +1,1 @@
+export { default as createTestAjv } from './validation/createTestAjv.js';
